[REFACTOR] User 엔티티, DTO 패키지 위치 변경
