+++ src/main/java/com/example/whiplash/domain/entity/UserAlarmInfo.java
@@ -1,6 +1,6 @@
 package com.example.whiplash.domain.entity;
 
-import com.example.whiplash.user.User;
+import com.example.whiplash.user.domain.User;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDateTime;
