+++ src/main/java/com/example/whiplash/domain/entity/history/email/EmailSendHistory.java
@@ -1,6 +1,6 @@
 package com.example.whiplash.domain.entity.history.email;
 
-import com.example.whiplash.user.User;
+import com.example.whiplash.user.domain.User;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDateTime;
