+++ src/main/java/com/example/whiplash/domain/entity/history/search/SearchHistory.java
@@ -1,6 +1,6 @@
 package com.example.whiplash.domain.entity.history.search;
 
-import com.example.whiplash.user.User;
+import com.example.whiplash.user.domain.User;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDateTime;
