+++ src/main/java/com/example/whiplash/domain/entity/trade/MockTradeLog.java
@@ -1,7 +1,7 @@
 package com.example.whiplash.domain.entity.trade;
 
 
-import com.example.whiplash.user.User;
+import com.example.whiplash.user.domain.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
