+++ src/main/java/com/example/whiplash/domain/repository/InvestorProfileRepository.java
@@ -1,10 +1,9 @@
 package com.example.whiplash.domain.repository;
 
-import com.example.whiplash.domain.entity.profile.InvestorProfile;
-import com.example.whiplash.user.User;
+import com.example.whiplash.user.domain.profile.InvestorProfile;
+import com.example.whiplash.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface InvestorProfileRepository extends JpaRepository<InvestorProfile, Long> {
