+++ src/main/java/com/example/whiplash/domain/repository/UserKeywordRepository.java
@@ -1,7 +1,7 @@
 package com.example.whiplash.domain.repository;
 
 import com.example.whiplash.domain.entity.UserKeyword;
-import com.example.whiplash.user.User;
+import com.example.whiplash.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
