+++ src/main/java/com/example/whiplash/user/domain/SocialProvider.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.user;
+package com.example.whiplash.user.domain;
 
 public enum SocialProvider {
     KAKAO
