+++ src/main/java/com/example/whiplash/user/domain/UserStatus.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.user;
+package com.example.whiplash.user.domain;
 
 public enum UserStatus {
     PENDING, ACTIVE, INACTIVE
