+++ src/main/java/com/example/whiplash/user/domain/profile/AgeRange.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.domain.entity.profile;
+package com.example.whiplash.user.domain.profile;
 
 public enum AgeRange {
     TWENTIES, THIRTIES, FOURTIES, FIFTIES, SIXTY_PLUS
