+++ src/main/java/com/example/whiplash/user/domain/profile/InvestmentGoal.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.domain.entity.profile;
+package com.example.whiplash.user.domain.profile;
 
 public enum InvestmentGoal {
     LONG_TERM_GROWTH, SHORT_TERM_PROFIT, STABLE_INCOME,
