+++ src/main/java/com/example/whiplash/user/domain/profile/RiskTolerance.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.domain.entity.profile;
+package com.example.whiplash.user.domain.profile;
 
 public enum RiskTolerance {
     STABLE, MODERATE, AGGRESSIVE
