+++ src/main/java/com/example/whiplash/user/web/dto/request/LoginRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.user.dto;
+package com.example.whiplash.user.web.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
