+++ src/main/java/com/example/whiplash/user/web/dto/request/ProfileRegisterDTO.java
@@ -1,12 +1,13 @@
-package com.example.whiplash.user.dto;
+package com.example.whiplash.user.web.dto.request;
 
-import com.example.whiplash.domain.entity.profile.AgeRange;
-import com.example.whiplash.domain.entity.profile.InvestmentGoal;
-import com.example.whiplash.domain.entity.profile.InvestmentLevel;
-import com.example.whiplash.domain.entity.profile.RiskTolerance;
+import com.example.whiplash.user.domain.profile.AgeRange;
+import com.example.whiplash.user.domain.profile.InvestmentGoal;
+import com.example.whiplash.user.domain.profile.InvestmentLevel;
+import com.example.whiplash.user.domain.profile.RiskTolerance;
 import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
 import lombok.Getter;
-
+@Builder
 @Getter
 public class ProfileRegisterDTO {
     @NotNull(message = "연령대는 필수입니다")
